+++ ai-logic/firebase-ai/src/testUtil/java/com/google/firebase/ai/JavaCompileTests.java
@@ -34,6 +34,7 @@ import com.google.firebase.ai.java.LiveModelFutures;
 import com.google.firebase.ai.java.LiveSessionFutures;
 import com.google.firebase.ai.java.TemplateChatFutures;
 import com.google.firebase.ai.java.TemplateGenerativeModelFutures;
+import com.google.firebase.ai.type.ActivityDetectionConfig;
 import com.google.firebase.ai.type.AspectRatio;
 import com.google.firebase.ai.type.BlockReason;
 import com.google.firebase.ai.type.Candidate;
@@ -75,6 +76,7 @@ import com.google.firebase.ai.type.MultiSpeakerVoiceConfig;
 import com.google.firebase.ai.type.Part;
 import com.google.firebase.ai.type.PromptFeedback;
 import com.google.firebase.ai.type.PublicPreviewAPI;
+import com.google.firebase.ai.type.RealtimeInputConfig;
 import com.google.firebase.ai.type.ResponseModality;
 import com.google.firebase.ai.type.RetrievalConfig;
 import com.google.firebase.ai.type.SafetyRating;
@@ -193,6 +195,18 @@ public class JavaCompileTests {
         .setPresencePenalty(2.0F)
         .setResponseModality(ResponseModality.AUDIO)
         .setSpeechConfig(new SpeechConfig(new Voice("AOEDE")))
+        .setRealtimeInputConfig(
+            new RealtimeInputConfig.Builder()
+                .setActivityHandling(RealtimeInputConfig.ActivityHandling.NO_INTERRUPT)
+                .setTurnCoverage(RealtimeInputConfig.TurnCoverage.ONLY_ACTIVITY)
+                .setAutomaticActivityDetection(
+                    new ActivityDetectionConfig.Builder()
+                        .setStartSensitivity(ActivityDetectionConfig.Sensitivity.HIGH)
+                        .setEndSensitivity(ActivityDetectionConfig.Sensitivity.LOW)
+                        .setPrefixPaddingMs(100)
+                        .setSilenceDurationMs(500)
+                        .build())
+                .build())
         .build();
   }
 
@@ -470,6 +484,8 @@ public class JavaCompileTests {
     session.sendAudioRealtime(new InlineData(bytes, "audio/jxl", null));
     session.sendVideoRealtime(new InlineData(bytes, "image/jxl", null));
     session.sendTextRealtime("text");
+    session.sendStartActivityRealtime();
+    session.sendStopActivityRealtime();
 
     FunctionResponsePart functionResponse =
         new FunctionResponsePart("myFunction", new JsonObject(Map.of()));